Refactor param replacement-related functions into own module
