+++ src/cwe_checker_lib/src/checkers/cwe_119/context/tests.rs
@@ -48,91 +48,3 @@ fn test_compute_size_value_of_malloc_like_call() {
     )
     .is_none());
 }
-
-#[test]
-fn test_substitute_param_values_context_sensitive() {
-    let mut context = Context::mock_x64();
-    let param_id = AbstractIdentifier::mock("func", "RDI", 8);
-    let callsite_id = AbstractIdentifier::mock("callsite_id", "RDI", 8);
-
-    let recursive_param_id = AbstractIdentifier::mock("main", "RSI", 8);
-    let recursive_callsite_id = AbstractIdentifier::mock("recursive_callsite_id", "RSI", 8);
-
-    let param_value = Data::from_target(recursive_param_id.clone(), Bitvector::from_i64(1).into());
-    let recursive_param_value = Data::from(Bitvector::from_i64(41));
-    let param_replacement_map = HashMap::from([
-        (callsite_id, param_value.clone()),
-        (recursive_callsite_id.clone(), recursive_param_value),
-    ]);
-    let callee_to_callsites_map = HashMap::from([
-        (Tid::new("func"), HashSet::from([Tid::new("callsite_id")])),
-        (
-            Tid::new("main"),
-            HashSet::from([Tid::new("recursive_callsite_id")]),
-        ),
-    ]);
-    let call_to_caller_map = HashMap::from([
-        (Tid::new("callsite_id"), Tid::new("main")),
-        (
-            Tid::new("recursive_callsite_id"),
-            Tid::new("somer_other_fn_id"),
-        ),
-    ]);
-    context.param_replacement_map = param_replacement_map;
-    context.callee_to_callsites_map = callee_to_callsites_map;
-    context.call_to_caller_fn_map = call_to_caller_map;
-    // non-recursive substitution
-    let result = context.substitute_param_values_context_sensitive(
-        &Data::from_target(param_id.clone(), Bitvector::from_i64(5).into()),
-        &Tid::new("callsite_id"),
-        &Tid::new("func"),
-    );
-    assert_eq!(
-        result,
-        Data::from_target(recursive_param_id.clone(), Bitvector::from_i64(6).into())
-    );
-    // recursive substitution
-    let result = context.recursively_substitute_param_values_context_sensitive(
-        &Data::from_target(param_id, Bitvector::from_i64(5).into()),
-        &Tid::new("func"),
-        &[Tid::new("callsite_id"), Tid::new("recursive_callsite_id")],
-    );
-    println!("{:#}", result.to_json_compact());
-    assert_eq!(result, Bitvector::from_i64(47).into());
-}
-
-#[test]
-fn test_substitute_param_values() {
-    let mut context = Context::mock_x64();
-    let param_id = AbstractIdentifier::mock("func", "RDI", 8);
-    let callsite_id = AbstractIdentifier::mock("callsite_id", "RDI", 8);
-
-    let recursive_param_id = AbstractIdentifier::mock("main", "RSI", 8);
-    let recursive_callsite_id = AbstractIdentifier::mock("recursive_callsite_id", "RSI", 8);
-
-    let param_value = Data::from_target(recursive_param_id.clone(), Bitvector::from_i64(1).into());
-    let recursive_param_value = Data::from(Bitvector::from_i64(39));
-    let param_replacement_map = HashMap::from([
-        (callsite_id, param_value.clone()),
-        (recursive_callsite_id.clone(), recursive_param_value),
-    ]);
-    let callee_to_callsites_map = HashMap::from([
-        (Tid::new("func"), HashSet::from([Tid::new("callsite_id")])),
-        (
-            Tid::new("main"),
-            HashSet::from([Tid::new("recursive_callsite_id")]),
-        ),
-    ]);
-    context.param_replacement_map = param_replacement_map;
-    context.callee_to_callsites_map = callee_to_callsites_map;
-    // non-recursive substitution
-    let (result_absolute, result) = context.substitute_param_values(&param_id);
-    assert!(result_absolute.is_none());
-    assert_eq!(result, param_value);
-    // recursive substitution
-    let result = context.recursively_substitute_param_values(&Data::from_target(
-        param_id,
-        Bitvector::from_i64(5).into(),
-    ));
-    assert_eq!(result, Bitvector::from_i64(45).into());
-}
